2/9

+++ NCKH/Core/Inc/DHT11.h
@@ -14,8 +14,8 @@ typedef struct {
 	TIM_HandleTypeDef *DHT_TIM;
 	uint16_t DHT_PIN;
 	GPIO_TypeDef *DHT_PORT;
-    float Temperature;
-    float Humidity;
+    uint16_t Temperature;
+    uint16_t Humidity;
 } DHT11_DataTypedef;
 
 void DHT11_Init(DHT11_DataTypedef *DHT, TIM_HandleTypeDef *Timer, GPIO_TypeDef *DHT_Port, uint16_t DHT_Pin);
